夏资陈电商平台 6.14 库存售罄

+++ coral-shop-core/src/main/java/com/dookay/coral/shop/goods/domain/PrototypeSpecificationOptionDomain.java
@@ -40,6 +40,6 @@ public class PrototypeSpecificationOptionDomain implements Serializable {
 
 	/*库存*/
 	@Transient
-	private int stock;
+	private Long stock;
 
 }
